Remove bundler app and move remnants to classifier.
Since we no longer will have any bundler views, the only things worth
saving are the model and serializer, and those don't really warrant
their own app without views.
