--- normandy/bundler/urls.py
@@ -1,8 +1,0 @@
-from django.conf.urls import url
-
-from normandy.bundler import views
-
-
-urlpatterns = [
-    url(r'^bundle/$', views.BundlerView.as_view(), name='normandy.bundler'),
-]
