+++ normandy/classifier/serializers.py
