+++ normandy/urls.py
@@ -7,7 +7,6 @@ from django.contrib import admin
 urlpatterns = [
     url(r'^admin/', admin.site.urls),
 
-    url(r'', include('normandy.bundler.urls')),
     url(r'', include('normandy.classifier.urls')),
     url(r'', include('normandy.selfrepair.urls')),
 ]
